Simplify scheduling by introducing an init process
Scheduling is easier if there is an init process which does nothing.
This way we don't have to perform hacks in order to idle in S privilege
mode while touching special registers in weird ways.

For now this process is kind of costly since 'wfi' is not accepted in
this context, but this can be further tuned down in the future by
setting TW=1 on 'mstatus', or by handling the exception and allowing it
if it comes from the proper process.

All of that being said, there is still work to be done as things fail
when optimitzations are on.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>