tests: Only test initrd addresses on debug
In initrd tests, and in general, we only need to check that the function
did something. As in, that it set the entry addresses somewhere and did
not leave them NULL. This is going to be performed from now on on CI and
regular tests. And, just if you want to go the extra mile, we will run
specific tests that check on memory addresses but that are not
guaranteed to work.

This is to be replaced in the future by having some proper debugging
output. As in, having a message at kernel boot time specifying the
addresses being used among other info. This needs a more complex printk
utility, of course.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>