Show the detected machine model
Originally I was planning to detect the machine model just in case we
needed to do workaround for special cases. But since apparently even
VisionFive2 brings its own CPU frequency base on DT, this is not even
needed. Hence, fetching the model is now a cool message being shown on
boot.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>