Allow the kernel to run as a Linux image
In order to run on real hardware, it's actually easier to re-use
workflows from existing bootloaders than loading things in memory
manually. In order to achieve this two things had to be settled:

1. The image has to have a Linux header as defined in the RISC-V port.
   This header will be taken into consideration by bootloaders in order to
   know where to jump, the image offset, the endianness of the image, etc.
2. The image cannot just be an ELF executable. Rather, we must translate
   it into binary form via `objcopy`.

With all of this at hand, I was able to make the kernel run on a
Starfive VisionFive 2 board, and I have recorded an example so it's
available as documentation.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>