Share the same stack everywhere
As documented in the code, we are not implementing any kind of memory
protection, so in theory any process (regardless if running in user or
kernel space) would be able to tamper with other processes' stack.
Hence, don't even pretend that we are separating stacks and share the
same global stack everywhere. This simplifies things a bit.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>