Simplify how we fetch the load address
Similarly to 58fcecc381be ("Simplify the constant on the size of the
stack"), the base and offset load addresses where taken from the Linux
kernel, which has to account for way more stuff than this poor kernel.

Hence, let's be more explicit about the base address and the offset
where the kernel will be located, so it's actually more clear both on
the linker, and on the kernel's header.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>