Re-use the halting loop also for the main hart
The halting in secondary harts should not be treated too differently as
with the (unlikely) scenario that the main function quits. Hence, merge
the code in both cases.

Moreover, we do not really need to to initialize the values of most
registers, as they will be explicitely set when needed anyways.

Signed-off-by: Miquel Sabaté Solà <mikisabate@gmail.com>